fix textarea and input
