+++ examples/src/doc/pages/forms/textarea.md
@@ -12,5 +12,6 @@ placeholder | 规定帮助用户填写输入字段的提示 | String | -
 readonly | 把输入字段设置为只读 | Boolean | false
 maxlength | 规定输入字段中的字符的最大长度 | Number | -
 autofocus | 规定输入字段在页面加载时是否获得焦点 | Boolean | false
+theme | 设置主题,可选参数（default，transparent，underline） | String | default
 
 ##### Textarea自动根据内容适配高度
