+++ src/mixins/input.js
@@ -1,7 +1,21 @@
 import Model from './model';
+import Config from '../config';
+
+const INPUT_PRE_STYLES = {
+    'default': 'background: #F3F6FB',
+    'transparent': 'background: transparent',
+    'underline': 'border-bottom: 1px solid #333',
+    ...Config('input.pre-themes')
+};
 
 export default {
     mixins: [Model],
+
+    data () {
+        return {
+            style: INPUT_PRE_STYLES[this.theme]
+        };
+    },
 
     props: {
         placeholder: {
@@ -34,7 +48,7 @@ export default {
             default: 'default'
         }
     },
-    
+
     methods: {
         onFocus () {
             !this.readonly && this.$emit('focus');
